Implement day 11
Unfortunately it is very slow.
I may see if I can speed it up before going on.

I suspect that it is because bigger square sizes are (probably)
exponentially slower (until they become so big that you start having
less of them).

There are 297*297 3 by 3s. This comes to 793,881 cells to add together.
There are 200*200 100 by 100s. This comes to 400,000,000 cells to add
together.

I suspect the approach I will take to fix this will be to not recalculate
every square, but work out what gets subtracted from it, and what gets
added to it.