Implement part 1 of day 14
I had a more generic digit iterator (beyond numbers between 0 and 19)
Since it generated least significant first, they'd be flipped in the
recipies vec. But for some reason this gave the wrong answer in the 2018
test case and the puzzle input.
I removed it because:
1. it only worked for some test cases
2. it was a lot more complexity than was required

I'm intrigued to know why it didn't work though.